Cleanup and API exposing
