+++ plugins/builtin/CMakeLists.txt
@@ -11,6 +11,7 @@ add_library(${PROJECT_NAME} SHARED
         source/content/lang_builtin_functions.cpp
         source/content/settings_entries.cpp
         source/content/tools_entries.cpp
+        source/content/data_processor_nodes.cpp
 
         source/math_evaluator.cpp
 )
