+++ plugins/libimhex/include/hex/helpers/shared_data.hpp
@@ -24,6 +24,7 @@ namespace hex::plugin::internal {
 namespace hex {
 
     namespace prv { class Provider; }
+    namespace dp { class Node; }
     class View;
 
     class SharedData {
@@ -66,6 +67,9 @@ namespace hex {
         static std::string fileBrowserTitle;
         static std::string fileBrowserValidExtensions;
         static std::function<void(std::string)> fileBrowserCallback;
+
+        static std::map<std::string, std::function<dp::Node*()>> dataProcessorNodes;
+        static u32 dataProcessorNodeIdCounter;
 
         static int mainArgc;
         static char **mainArgv;
